refactor: fold starter generator and create-heyapollo into apps/wizard
tooling/ is gone: apps/wizard is now the create-heyapollo package —
flat src/ stays the wizard shipped as setup/, cli/ is the scaffolder
bin, generator/ builds the embedded starter template. Tests merge
under __tests__/, workflows and docs follow the new paths, and the
npm README now leads with the branding banner and a brain-only note
pointing at the firmware-adaptation skills.
