+++ apps/wizard/.gitignore
@@ -1,2 +1,3 @@
 dist/
 template/
+generator/out/
