+++ apps/wizard/README.md
@@ -7,7 +7,7 @@
 
 <br />
 <div align="center">
-  <img src="https://raw.githubusercontent.com/galfrevn/apollo/main/.github/assets/banner.jpeg" alt="Apollo banner" width="100%">
+  <img src="https://raw.githubusercontent.com/galfrevn/apollo/main/branding/banner.png" alt="Apollo banner" width="100%">
 
   <h3 align="center">create-heyapollo</h3>
 
@@ -35,6 +35,10 @@ npm create heyapollo
 One command copies the Apollo starter into `apollo/` (or a name you pass), initializes git, installs dependencies, and drops you into the interactive setup wizard: Cloudflare account confirmation, live API key validation, an ElevenLabs voice picker, your city and timezone, then provision, deploy, and a verified device handshake. "No keys yet" is a first class answer: trial mode deploys a fully protocol correct brain with zero external spend.
 
 The scaffolded project requires [Bun](https://bun.sh). The Cloudflare Workers free plan is enough for everything except the optional coding sandbox.
+
+## Just the brain
+
+What you scaffold is the brain alone — the Cloudflare Worker that listens, thinks, speaks, and remembers. No firmware is included. Instead, the project ships agent skills (`.claude/skills`) that document Apollo's wire protocol end to end, so your coding agent can adapt the firmware of your own device to it — the [reference firmware](https://github.com/galfrevn/apollo-firmware) shows a complete ESP32 implementation.
 
 ## Flags
