+++ apps/wizard/generator/assets/CLAUDE.md
