+++ apps/wizard/generator/assets/scripts/bootstrap.ts
