+++ apps/wizard/generator/assets/scripts/flags.ts
