+++ apps/wizard/generator/assets/scripts/messages.ts
