+++ apps/wizard/generator/assets/scripts/probe.ts
