+++ apps/wizard/generator/assets/scripts/provision.ts
