+++ apps/wizard/generator/assets/scripts/shell.ts
