+++ apps/wizard/generator/assets/scripts/vars.ts
