--- tooling/create/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "@apollo/typescript-config/script.json",
-  "compilerOptions": {
-    "paths": {
-      "@/*": ["./src/*"]
-    }
-  },
-  "include": ["*.ts", "src/**/*.ts", "__tests__/*.ts"]
-}
