--- tooling/starter/.gitignore
@@ -1,1 +1,0 @@
-out/
