--- tooling/starter/bunfig.toml
@@ -1,2 +1,0 @@
-[test]
-root = "./__tests__"
