--- tooling/starter/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "@apollo/starter",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "generate": "bun build.ts",
-    "smoke": "bun build.ts --smoke",
-    "typecheck": "tsc --noEmit",
-    "test": "bun test"
-  },
-  "devDependencies": {
-    "@apollo/typescript-config": "workspace:*",
-    "typescript": "^5.9.3",
-    "zod": "^3.25.76"
-  }
-}
