--- tooling/starter/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "@apollo/typescript-config/script.json",
-  "compilerOptions": {
-    "paths": {
-      "@/*": ["./*"]
-    }
-  },
-  "include": ["*.ts", "assets/scripts/*.ts", "__tests__/*.ts"]
-}
